Create HomePage
In this commit I created HomePage. I created layout in main folder for our main pages. It contains the header, children and the footer. Header component is not yet finished but at least the structure of it is finished. Header as well as whole HomePage is responsive on mobile. MobileModal component and useMobileModal hook helps achieve this. Two new fonts are added. There some of the components created to separate the logic and make code more readable. One of the components is Button component which we can now reuse in our whole project for consistency.