Implement user authentication functionality
In this commit NextAuth is added to the project. User is now able to sign in with Google and users data is being stored in the database. In the header the icon to sign in will be shown. If the user is already signed it then the user profile icon will appear that will act as dropdown with link to profile page and sign out. On mobile this dropdown will tkae whole screen because of our mobile modal functionality. Also dropdown is reusable and in future will be used for bag/cart dropdown.