Create clothing item and wishlist page
In this commit I added clothing item page that shows information about the item. In this page similar items are also shown. User can select desired color and size of the item. User an also save the item to the wishlist. Items that user save will be shown in wishlist page. Wishlist table is added to the database. More clothing items were added to be seeded in the database.