Implement bag functionality and payments with Stripe
User can now add and remove clothing items to the bag. If same item is added twice the quantity will increase instead of adding new item. User cannot add the item until color and size are chosen. User also cannot add item if it is out of stock. When at least one item is added to the bag the checkout button will appear. Payments are done using Stripe. If the payment is successful quantity of purchased items will be removed from the database. User will be navigated to the payment page showing the status of the payment regarding if the payment is successful or canceled. On the profile page, users payment history is shown.