Add Stripe session check, profile icons links bigger and refactor code
In this commit I created interface for props for all components. I renamed Prisma models all to be singular for consistency and updated types. Profile icon links are now bigger and easier to click on. Updated zustand state callback to be more performant. When handling Stripe sessions we add them to the database and check if it is already processed so that we don't decrease quantity more then needed. This check can be used for more then just updating the quantity.